Fixes dark mode initial loading state and toggle
