+++ forgettable-frontend/src/App.js
@@ -28,6 +28,13 @@ function App() {
       }
       setIsLoggingIn(false);
     });
+
+    const theme = localStorage.getItem('theme');
+    if (theme == 'dark') {
+      document.documentElement.setAttribute('data-theme', 'dark');
+    } else {
+      document.documentElement.setAttribute('data-theme', 'light');
+    }
   }, []);
 
   const login = () => {
